[http-client] Only copy CA certs in build.rs if not already present.
This resolves an issue present on Arch Linux (at least) where the CA
certs file that is found and copied has permissions of 444 (i.e.
read-only). When the build script is compiled and re-run it is trying to
copy over a read-only file and triggered a build failure with
"Permission denied". This fix skips a re-copying of the same file--less
work to do, why not?

+++ components/http-client/build.rs
@@ -11,7 +11,9 @@ mod inner {
     pub fn main() {
         let src = env::var("SSL_CERT_FILE").unwrap();
         let dst = Path::new(&env::var("OUT_DIR").unwrap()).join("cacert.pem");
-        fs::copy(src, dst).unwrap();
+        if !dst.exists() {
+            fs::copy(src, dst).unwrap();
+        }
     }
 }
